Fixed readme and lamp startup script

+++ run-lamp.sh
@@ -58,16 +58,19 @@ fi
 # Run Postfix
 /usr/sbin/postfix start
 
+if [ ! -f /root/mysql_defaults.sql ]; then
+    # Run MariaDB (to set root pw)
+    /usr/bin/mysqld_safe --skip-grant-tables
+
+    # Set MariaDB/MySQL defaults
+    mysql < /root/mysql_defaults.sql
+
+    service mysql stop
+    rm /root/mysql_defaults.sql
+fi
+
 # Run MariaDB
-/usr/bin/mysqld_safe --skip-grant-tables --timezone=${DATE_TIMEZONE}&
-
-# Set MariaDB/MySQL defaults
-mysql < /root/mysql_defaults.sql
-
-service mysql stop
-service mysql start
-
-rm /root/mysql_defaults.sql
+service mysql restart
 
 # Run Apache:
 if [ $LOG_LEVEL == 'debug' ]; then
@@ -75,3 +78,5 @@ if [ $LOG_LEVEL == 'debug' ]; then
 else
     &>/dev/null /usr/sbin/apachectl -DFOREGROUND -k start
 fi
+
+echo "DONE"
